+++ src/app/shared/FeedFormSectionKinds.svelte
@@ -58,8 +58,6 @@
     {label: "Image", kind: FILE_METADATA},
     {label: "Relay selections", kind: RELAYS},
   ])
-
-  console.log(helper)
 </script>
 
 <span class="staatliches text-lg">What kind of content do you want to see?</span>