+++ src/app/shared/NoteContentKind1063.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import {Tags} from "@welshman/util"
-  import {urlIsMedia} from "src/util/notes"
+  import {urlIsMedia} from "@welshman/content"
   import NoteContentLink from "src/app/shared/NoteContentLink.svelte"
 
   export let note, showMedia