+++ src/app/shared/NoteContentKind1808.svelte
@@ -4,13 +4,13 @@
   import NoteContentLabel from "src/app/shared/NoteContentLabel.svelte"
   import NoteContentKind1 from "src/app/shared/NoteContentKind1.svelte"
 
-  export let note, maxLength, showEntire
+  export let note, showEntire
 
   const {stream_url} = fromPairs(note.tags)
 </script>
 
 <div class="flex flex-col gap-2 overflow-hidden text-ellipsis">
-  <NoteContentKind1 {note} {maxLength} {showEntire} />
+  <NoteContentKind1 {note} {showEntire} />
   {#if stream_url}
     <Media url={stream_url} />
   {/if}