+++ src/app/shared/NoteContentKind1986.svelte
@@ -3,12 +3,12 @@
   import NoteContentRating from "src/app/shared/NoteContentRating.svelte"
   import NoteContentKind1 from "src/app/shared/NoteContentKind1.svelte"
 
-  export let note, maxLength, showEntire
+  export let note, showEntire
 
   const rating = getRating(note)
 </script>
 
 <div class="flex flex-col gap-2 overflow-hidden text-ellipsis">
   <NoteContentRating {note} {rating} />
-  <NoteContentKind1 {note} {maxLength} {showEntire} />
+  <NoteContentKind1 {note} {showEntire} />
 </div>